# docs/swagger.yaml
definitions:
...
  util.CommonResult:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
...
  /resourceCategory/create:
    post:
      consumes:
      - application/json
      description: 创建资源分类
      parameters:
      - description: 创建时间
        in: query
        name: createTime
        type: string
      - in: query
        name: id
        type: integer
      - description: 分类名称
        in: query
        name: name
        type: string
      - description: 排序
        in: query
        name: sort
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.CommonResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.CommonResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.CommonResult'
      summary: 创建资源分类
      tags:
      - 后台资源分类管理
  /resourceCategory/delete/{id}:
    post:
      consumes:
      - application/json
      description: 修改资源分类
      parameters:
      - description: id
        in: path
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.CommonResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.CommonResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.CommonResult'
      summary: 修改资源分类
      tags:
      - 后台资源分类管理
  /resourceCategory/listAll:
    get:
      consumes:
      - application/json
      description: 资源分类列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.CommonResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.CommonResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.CommonResult'
      summary: 资源分类列表
      tags:
      - 后台资源分类管理
  /resourceCategory/update/{id}:
    post:
      consumes:
      - application/json
      description: 修改资源分类
      parameters:
      - description: id
        in: path
        name: id
        type: string
      - description: 创建时间
        in: query
        name: createTime
        type: string
      - in: query
        name: id
        type: integer
      - description: 分类名称
        in: query
        name: name
        type: string
      - description: 排序
        in: query
        name: sort
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.CommonResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.CommonResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.CommonResult'
      summary: 修改资源分类
      tags:
      - 后台资源分类管理
...
  /role/allocMenu:
    post:
      consumes:
      - application/json
      description: 菜单分配
      parameters:
      - description: roleId
        in: query
        name: roleId
        type: string
      - description: menuIds
        in: query
        items:
          type: string
        name: menuIds
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.CommonResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.CommonResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.CommonResult'
      summary: 菜单分配
      tags:
      - 后台用户角色管理
  /role/allocResource:
    post:
      consumes:
      - application/json
      description: 资源分配
      parameters:
      - description: roleId
        in: query
        name: roleId
        type: string
      - description: resourceIds
        in: query
        items:
          type: string
        name: resourceIds
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.CommonResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.CommonResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.CommonResult'
      summary: 资源分配
      tags:
      - 后台用户角色管理
  /role/create:
    post:
      consumes:
      - application/json
      description: 创建角色
      parameters:
      - description: 后台用户数量
        in: query
        name: adminCount
        type: integer
      - description: 创建时间
        in: query
        name: createTime
        type: string
      - description: 描述
        in: query
        name: description
        type: string
      - in: query
        name: id
        type: integer
      - description: 名称
        in: query
        name: name
        type: string
      - in: query
        name: sort
        type: integer
      - description: 启用状态：0->禁用；1->启用
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.CommonResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.CommonResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.CommonResult'
      summary: 创建角色
      tags:
      - 后台用户角色管理
  /role/delete/{id}:
    post:
      consumes:
      - application/json
      description: 删除角色
      parameters:
      - description: id
        in: path
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.CommonResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.CommonResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.CommonResult'
      summary: 删除角色
      tags:
      - 后台用户角色管理
  /role/list:
    get:
      consumes:
      - application/json
      description: 角色列表
      parameters:
      - description: 名称
        in: query
        name: keyword
        type: string
      - description: 页码
        in: query
        name: pageNum
        type: string
      - description: 数量
        in: query
        name: pageSize
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.CommonResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.CommonResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.CommonResult'
      summary: 角色列表
      tags:
      - 后台用户角色管理
  /role/listAll:
    get:
      consumes:
      - application/json
      description: 获取所有角色
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.CommonResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.CommonResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.CommonResult'
      summary: 获取所有角色
      tags:
      - 后台用户角色管理
  /role/listMenu/{roleId}:
    get:
      consumes:
      - application/json
      description: 根据角色id获取菜单
      parameters:
      - description: roleId
        in: path
        name: roleId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.CommonResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.CommonResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.CommonResult'
      summary: 根据角色id获取菜单
      tags:
      - 后台用户角色管理
  /role/listResource/{roleId}:
    get:
      consumes:
      - application/json
      description: 根据角色id获取资源
      parameters:
      - description: roleId
        in: path
        name: roleId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.CommonResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.CommonResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.CommonResult'
      summary: 根据角色id获取资源
      tags:
      - 后台用户角色管理
  /role/update/{id}:
    post:
      consumes:
      - application/json
      description: 修改角色
      parameters:
      - description: id
        in: path
        name: id
        type: string
      - description: 后台用户数量
        in: query
        name: adminCount
        type: integer
      - description: 创建时间
        in: query
        name: createTime
        type: string
      - description: 描述
        in: query
        name: description
        type: string
      - in: query
        name: id
        type: integer
      - description: 名称
        in: query
        name: name
        type: string
      - in: query
        name: sort
        type: integer
      - description: 启用状态：0->禁用；1->启用
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.CommonResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.CommonResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.CommonResult'
      summary: 修改角色
      tags:
      - 后台用户角色管理
  /role/updateStatus/{id}:
    post:
      consumes:
      - application/json
      description: 修改角色状态
      parameters:
      - description: id
        in: path
        name: id
        type: string
      - description: status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.CommonResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.CommonResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.CommonResult'
      summary: 修改角色状态
      tags:
      - 后台用户角色管理
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"